pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sreality-scraper"
version = "0.1.0"
description = "Scraper for sreality.cz Czech real estate portal"
requires-python = ">=3.12"

dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "httpx>=0.27",
    "parsel>=1.9",
    "playwright>=1.44",
    "apscheduler>=3.10",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "black>=24.0",
    "isort>=5.13",
    "mypy>=1.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
